refactor css files into styles folder
