+++ src/App.js
@@ -9,8 +9,8 @@ function App() {
 			<Switch>
 				<>
 					<Route exact path="/" component={HomePage} />
-					<Route path="/games" component = {GamesListContainer}/>
-					<Route path="/scores" component = {ScoreListContainer}/>
+					<Route path="/games" component={GamesListContainer} />
+					<Route path="/scores" component={ScoreListContainer} />
 				</>
 			</Switch>
 		</Router>
