+++ src/components/Header.js
@@ -1,17 +1,16 @@
 import React from 'react';
 import NavBar from './NavBar';
-import './Header.css';
-import './NavBar.css';
+import '../styles/Header.css';
 
 function Header() {
 
 
-    return(
+    return (
         <>
-        <h1 className="header-title">Jade Arcade</h1>
-        <ul className="navbar">
-           <NavBar />
-        </ul>
+            <h1 className="header-title">Jade Arcade</h1>
+            <ul className="navbar">
+                <NavBar />
+            </ul>
         </>
     )
 };
