+++ src/components/HomePage.js
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import './HomePage.css'
+import '../styles/HomePage.css';
 
 function HomePage() {
     return (
