+++ src/containers/GamesListContainer.js
@@ -1,8 +1,5 @@
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Header from '../components/Header';
 import Footer from '../components/Footer';
-import GameContainer from '../containers/GameContainer';
-import ScoreListContainer from '../containers/ScoreListContainer';
 
 function GamesListContainer() {
