+++ src/containers/ScoreListContainer.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Header from '../components/Header';
 import Footer from '../components/Footer';
 
@@ -7,9 +6,9 @@ function ScoreListContainer() {
 
     return (
         <>
-            <Header/>
+            <Header />
             <h2>Score list </h2>
-            <Footer/>
+            <Footer />
         </>
     )
 }
